+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Write job ID to log when re-queueing deliveries.
 - Use UUIDv7 instead of ULID for primary keys.
 - Added default values for `http_host` and `http_port` configuration parameters.
+- Delete actor profile if inbox endpoint returns `410 Gone`.
 
 ### Fixed